feat(run): run is async, no longer need resolvedTarget
Users were not expecting to need to pass in the optional resolvedTarget, and it would provide
warnings on some rules even when there were no $ref's in there. This will automatically resolve
local refs and dereference them too, which is then an object some rules can chose to look at instead
of the original object.

BREAKING CHANGE: The spectral.run() method is now async, so the await keyword must be prepended.

fix #88