refactor(oas): removed schema-items-is-object rule
This rule seems problematic in the way it was implemented, and does not seem particularly useful
anyway.

BREAKING CHANGE: If you were relying on schema-items-is-object to error on schemas that had this
invalid OpenAPI, then it will give you an error in a slightly different way now, but should not let
invalid OpenAPI files pass.