feat: merge openapi-tags and openapi-tags-alphabetical
Two very similar rules, both were not recommended but people were not aware they needed to enable tags for Stoplight Docs. Having tags is just generally a great idea and something I've always recommended in general, so this seems like a reasonable addition. Now that studio and custom rulesets are more in the forefront, I figure folks can always turn this off if they hate it. It is also now recommended, because not recommended things are basically completely hidden from most users.