fix: ruleset path for Example Object validation
The 'example' field has 'Any' values in all contexts
it can be used so the rule should not be triggered for that field.

The 'examples' field always has values of Example Object type in
all contexts, so this is what the rule should trigger on.

Changed tests to validate `examples` instead of `example` and added
additional tests with multiple examples below the `examples` field.