chore(doc): Make the xor function example is little less abstract
This example has been taken from the initial `example-value-or-externalValue`
definition in the ruleset. The rule has since been updated but this doco example
hasn't.

Let's make the doco example a litlle less dependent on the concrete rule implementation.
It's sole purpose is to provide a sample usage of the `xor`function.