fix(core): consider `message` when de-duplicating results
Before this change, the `code` and `path` of the results were the primary factors
considered in de-duplication. Results that came from the same rule and same OpenAPI
artifact but had different messages were de-duplicated and not all unique results
were returned.

This change updates the logic to take `message` into account so that all unique
results are displayed to the user.